+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=PATCH
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* The validation for Dashboard#uid field has been added.
+It can contain no more than 40 characters now, according to the Grafana API. Otherwise exception will be thrown.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [2.7.0]() (09-10-2019)